refactor(cli): extract dictate sub-app to cli_dictate.py (ADR-059 V9)
